requirements.txt
pdfplumber>=0.11.0
openpyxl>=3.1.0
thefuzz>=0.22.0
openai>=1.0.0
python-dotenv>=1.0.0
